+++ site/dat/docs/Changelog.md
@@ -7,6 +7,7 @@
   - fix test duration logic to respect iterations set
   - distributed tests are now supported with JMeter UI (gui: true option)
   - install JMeter-Plugins 1.3.0
+  - all default tool locations are now under `~/.bzt` dir
 
 # 0.2.23
   - rename `bulk-size` option into `send-interval` for BlazeMeter reporter