+++ site/dat/docs/Changelog.md
@@ -9,6 +9,7 @@
   - install JMeter-Plugins 1.3.0
   - all default tool locations are now under `~/.bzt` dir
   - FinalStatus reporter now provides test duration (test-duration option, True by default)
+  - bzt now fails when no requests were given in scenario
 
 ## 0.2.23
   - rename `bulk-size` option into `send-interval` for BlazeMeter reporter